전략 수정 시 제안서 업로드 할 경우 기존 제안서 파일이 없는 전략일 경우에 대한 예외 처리

+++ src/main/java/com/investmetic/domain/strategy/service/StrategyService.java
@@ -301,7 +301,10 @@ public class StrategyService {
 
             String presignedUrl = s3FileService.getPreSignedUrl(proposalFilePath);
 
-            s3FileService.deleteFromS3(strategy.getProposalFilePath());
+            if (!strategy.getProposalFilePath().isEmpty()) {
+                s3FileService.deleteFromS3(strategy.getProposalFilePath());
+            }
+
             strategy.modifyStrategyWithProposalFilePath(requestDto.getStrategyName(), requestDto.getDescription(),
                     proposalFilePath);
